[++] changed to float16 for 3D nets [==] refactored a couple of things for 3D experiments

+++ Training/Utils/cnn_utils.py
@@ -168,7 +168,7 @@ def _conv_layer_pure_3d(input, shape, strides=[1, 1, 1, 1, 1], padding='SAME'):
 		'weights',
 		shape,
 		initializer=tf.contrib.layers.variance_scaling_initializer(),
-		dtype=tf.float32)
+		dtype=tf.float16)
 
 	conv = tf.nn.conv3d(input, kernel, strides=strides, padding=padding, name='Pre-Activation')
